move teleop class to its own file
